+++ src/Digipolis.Correlation/Constants/ErrorCode.cs
@@ -1,0 +1,8 @@
+﻿namespace Digipolis.Correlation
+{
+    internal static class ErrorCode
+    {
+        public const string InvalidCorrelationHeader = "INVCOR";
+        public const string RequiredCorrelationHeader = "REQCOR";
+    }
+}