+++ src/Digipolis.Correlation/Formatters/ICorrelationContextFormatter.cs
@@ -1,0 +1,7 @@
+﻿namespace Digipolis.Correlation
+{
+    public interface ICorrelationContextFormatter
+    {
+        CorrelationContext ValidateAndSetPropertiesFromDgpHeader(string header);
+    }
+}